\name{marbotte_util-package}
\alias{marbotte_util-package}
\alias{marbotte_util}
\docType{package}
\title{
\packageTitle{marbotte_util}
}
\description{
\packageDescription{marbotte_util}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{marbotte_util}
\packageIndices{marbotte_util}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{marbotte_util}

Maintainer: \packageMaintainer{marbotte_util}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
